requirements.txt
asgiref==3.8.1
Django==5.0.4
django-phonenumber-field==7.3.0
django-phonenumbers==1.0.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
gunicorn==22.0.0
inflection==0.5.1
packaging==24.0
phonenumbers==8.13.35
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
setuptools==69.5.1
sqlparse==0.5.0
uritemplate==4.1.1
